+++ e3nn_jax/_src/mlp_haiku.py
@@ -5,9 +5,6 @@ import jax.numpy as jnp
 import e3nn_jax as e3nn
 
 
-# TODO move to haiku file
-
-
 class MultiLayerPerceptron(hk.Module):
     """Just a simple MLP for scalars. No equivariance here.