+++ e3nn_jax/flax.py
@@ -1,3 +1,4 @@
 from e3nn_jax._src.linear_flax import Linear
+from e3nn_jax._src.mlp_flax import MultiLayerPerceptron
 
-__all__ = ["Linear"]
+__all__ = ["Linear", "MultiLayerPerceptron"]