+++ e3nn_jax/haiku.py
@@ -1,3 +1,4 @@
 from e3nn_jax._src.linear_haiku import Linear
+from e3nn_jax._src.mlp_haiku import MultiLayerPerceptron
 
-__all__ = ["Linear"]
+__all__ = ["Linear", "MultiLayerPerceptron"]